test(app): share disabled speech e2e env
